+++ CHANGELOG.md
@@ -1,4 +1,22 @@
 ## v1.7.16
+## v1.7.16
+
+### Bug Fixes
+
+1. [#5265](https://github.com/influxdata/chronograf/pull/5323): fix(schema-explorer): update the flux schema explorer to use v1 package
+1. [#5326](https://github.com/influxdata/chronograf/pull/5326): fix(ui): use a fallback label for y-axis if it's available
+1. [#5334](https://github.com/influxdata/chronograf/pull/5334): fix(influx): allow upperDashboardTime when generating query config
+1. [#5335](https://github.com/influxdata/chronograf/pull/5335): fix(kapacitor): use better heuristic than string contains `batch` when creating tasks
+
+### Features
+
+## v1.7.15
+
+### Features
+
+1. [5324](https://github.com/influxdata/chronograf/pull/5324): Pin to latest minor go version; make Docker build process more robust
+
+## v1.7.15
 
 ### Bug Fixes